+++ src/metworkpy/__init__.py
@@ -3,24 +3,34 @@ from importlib.metadata import version
 __author__ = "Braden Griebel"
 __version__ = version("metworkpy")
 __all__ = [
-    "utils",
-    "imat",
-    "gpr",
-    "information",
     "divergence",
+    "gpr",
+    "hyper",
+    "imat",
+    "information",
+    "metabolites",
     "network",
     "sampling",
     "synleth",
-    "read_model",
-    "write_model",
-    "model_eq",
-    "model_bounds_eq",
-    "mutual_information",
-    "variation_of_information",
-    "mi_network_adjacency_matrix",
+    "utils",
+    # Divergence
     "kl_divergence",
     "js_divergence",
     "get_example_model",
+    # Mutual information
+    "mutual_information",
+    "variation_of_information",
+    "mi_network_adjacency_matrix",
+    # Hyper
+    "HyperGraph",
+    # Metabolites
+    "metchange",
+    "find_metabolite_synthesis_network_genes",
+    "find_metabolite_synthesis_network_reactions",
+    # Networks
+    "bipartite_project",
+    "fuzzy_reaction_set",
+    "fuzzy_reaction_intersection",
     "create_metabolic_network",
     "create_reaction_network",
     "create_metabolite_network",
@@ -35,35 +45,34 @@ __all__ = [
     "create_adjacency_matrix",
     "node_target_density",
     "find_dense_clusters",
+    # Utils
+    "eval_gpr",
+    "gene_to_rxn_weights",
+    "fisher_enrichment",
+    "extended_mannwhitneyu_test",
     "reaction_to_gene_ids",
     "gene_to_reaction_ids",
     "gene_to_reaction_list",
     "reaction_to_gene_list",
     "get_gene_to_reaction_translation_dict",
     "get_reaction_to_gene_translation_dict",
-    "bipartite_project",
-    "fuzzy_reaction_set",
-    "fuzzy_reaction_intersection",
-    "find_metabolite_synthesis_network_genes",
-    "find_metabolite_synthesis_network_reactions",
-    "metchange",
-    "metabolites",
-    "eval_gpr",
-    "gene_to_rxn_weights",
-    "fisher_enrichment",
-    "extended_mannwhitneyu_test",
+    "read_model",
+    "write_model",
+    "model_eq",
+    "model_bounds_eq",
 ]
 
 from metworkpy import (
-    utils,
-    imat,
-    gpr,
-    information,
     divergence,
+    gpr,
+    hyper,
+    imat,
+    information,
     metabolites,
     network,
     sampling,
     synleth,
+    utils,
 )
 
 from metworkpy.utils import (
@@ -114,3 +123,5 @@ from metworkpy.metabolites import (
 )
 
 from metworkpy.gpr import eval_gpr, gene_to_rxn_weights
+
+from metworkpy.hyper import HyperGraph