+++ src/metworkpy/hyper/__init__.py
@@ -1,0 +1,8 @@
+"""
+Module defining hypergraphs, and
+functions to work with them
+"""
+
+__all__ = ["HyperGraph"]
+
+from .hypergraph import HyperGraph